Internal: Interactions - data handling [ED-21548]
 * extracted interactions list gen. from the module
