+++ modules/interactions/module.php
@@ -15,18 +15,18 @@ class Module extends BaseModule {
 	const MODULE_NAME = 'e-interactions';
 	const EXPERIMENT_NAME = 'e_interactions';
 
-	const TRIGGERS = [ 'load', 'scrollIn', 'scrollOut' ];
-	const EFFECTS = [ 'fade', 'slide', 'scale' ];
-	const TYPES = [ 'in', 'out' ];
-	const DIRECTIONS = [ 'left', 'right', 'top', 'bottom' ];
-	const DEFAULT_DURATION = 300;
-	const DEFAULT_DELAY = 0;
-	const SLIDE_DISTANCE = 100;
-	const SCALE_START = 0.5;
-	const EASING = 'linear';
-
 	public function get_name() {
 		return self::MODULE_NAME;
+	}
+
+	private $animations;
+
+	private function get_animations() {
+		if ( ! $this->animations ) {
+			$this->animations = new Animations();
+		}
+
+		return $this->animations;
 	}
 
 	public static function get_experimental_data() {
@@ -42,7 +42,7 @@ class Module extends BaseModule {
 
 	public function is_experiment_active() {
 		return Plugin::$instance->experiments->is_feature_active( self::EXPERIMENT_NAME )
-				&& Plugin::$instance->experiments->is_feature_active( AtomicWidgetsModule::EXPERIMENT_NAME );
+			&& Plugin::$instance->experiments->is_feature_active( AtomicWidgetsModule::EXPERIMENT_NAME );
 	}
 
 	public function __construct() {
@@ -57,90 +57,14 @@ class Module extends BaseModule {
 		add_action( 'elementor/frontend/before_enqueue_scripts', fn () => $this->enqueue_interactions() );
 		add_action( 'elementor/preview/enqueue_scripts', fn () => $this->enqueue_preview_scripts() );
 		add_action( 'elementor/editor/after_enqueue_scripts', fn () => $this->enqueue_editor_scripts() );
+
 		add_filter( 'elementor/document/save/data', fn ( $data ) => $this->sanitize_document_data( $data ), 10, 1 );
-	}
-
-	private function get_label( $key, $value ) {
-		$special_labels = [
-			'trigger' => [
-				'load' => __( 'On page load', 'elementor' ),
-				'scrollIn' => __( 'Scroll into view', 'elementor' ),
-				'scrollOut' => __( 'Scroll out of view', 'elementor' ),
-			],
-		];
-
-		if ( isset( $special_labels[ $key ][ $value ] ) ) {
-			return $special_labels[ $key ][ $value ];
-		}
-
-		$label = ucwords( str_replace( '-', ' ', $value ) );
-
-		return esc_html( $label );
-	}
-
-	private function generate_animation_options() {
-		$options = [];
-		$durations = [ 0, 100, 200, 300, 400, 500, 750, 1000, 1250, 1500 ];
-		$delays = [ 0, 100, 200, 300, 400, 500, 750, 1000, 1250, 1500 ];
-
-		foreach ( self::TRIGGERS as $trigger ) {
-			foreach ( self::EFFECTS as $effect ) {
-				foreach ( self::TYPES as $type ) {
-					foreach ( self::DIRECTIONS as $direction ) {
-						foreach ( $durations as $duration ) {
-							foreach ( $delays as $delay ) {
-								$value = "{$trigger}-{$effect}-{$type}-{$direction}-{$duration}-{$delay}";
-								$label = sprintf(
-									'%s: %s %s %s (%dms/%dms)',
-									$this->get_label( 'trigger', $trigger ),
-									$this->get_label( 'effect', $effect ),
-									$this->get_label( 'type', $type ),
-									$this->get_label( 'direction', $direction ),
-									$duration,
-									$delay
-								);
-								$options[] = [
-									'value' => $value,
-									'label' => $label,
-								];
-							}
-						}
-					}
-
-					foreach ( $durations as $duration ) {
-						foreach ( $delays as $delay ) {
-							$value = "{$trigger}-{$effect}-{$type}--{$duration}-{$delay}";
-							$label = sprintf(
-								'%s: %s %s (%dms/%dms)',
-								$this->get_label( 'trigger', $trigger ),
-								$this->get_label( 'effect', $effect ),
-								$this->get_label( 'type', $type ),
-								$duration,
-								$delay
-							);
-							$options[] = [
-								'value' => $value,
-								'label' => $label,
-							];
-						}
-					}
-				}
-			}
-		}
-
-		return $options;
 	}
 
 	private function get_config() {
 		return [
-			'constants' => [
-				'defaultDuration' => self::DEFAULT_DURATION,
-				'defaultDelay' => self::DEFAULT_DELAY,
-				'slideDistance' => self::SLIDE_DISTANCE,
-				'scaleStart' => self::SCALE_START,
-				'easing' => self::EASING,
-			],
-			'animationOptions' => $this->generate_animation_options(),
+			'constants' => $this->get_animations()->defaults(),
+			'animationOptions' => $this->get_animations()->list(),
 		];
 	}
 
@@ -205,7 +129,7 @@ class Module extends BaseModule {
 		static $valid_ids = null;
 
 		if ( null === $valid_ids ) {
-			$valid_ids = array_column( $this->generate_animation_options(), 'value' );
+			$valid_ids = array_column( $this->get_animations()->list(), 'value' );
 		}
 
 		if ( ! is_string( $animation_id ) || empty( $animation_id ) ) {
